git-svn-id: https://www.seasar.org/svn/s2struts/s2struts-backport175@29 c287a166-f701-0410-b539-9faeb104086e

+++ src/main/org/seasar/struts/validator/annotation/ByteType.java
@@ -3,6 +3,5 @@ package org.seasar.struts.validator.annotation;
 /**
  * @author Satoshi Kimura
  */
-public interface Float extends CommonValidator {
-
+public interface ByteType extends CommonValidator {
 }
