+++ src/main/org/seasar/struts/validator/annotation/ShortType.java
@@ -3,6 +3,6 @@ package org.seasar.struts.validator.annotation;
 /**
  * @author Satoshi Kimura
  */
-public interface Integer extends CommonValidator {
+public interface ShortType extends CommonValidator {
 
 }
